requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.6
django-allauth==65.3.0
django-allauth-ui==1.5.0
django-components==0.123
django-widget-tweaks==1.5.0
whitenoise==6.6.0
gunicorn==23.0.0
facenet_pytorch==2.6.0
fastai==1.0.60
fastprogress==1.0.3
ffmpeg==1.4
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
joblib==1.4.2
kornia==0.8.0
matplotlib==3.10.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
pandas==2.2.3
pillow==10.2.0
pycocotools==2.0.8
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
selectolax==0.3.26
six==1.17.0
slippers==0.6.2
sqlparse==0.5.2
tensorflow==2.14.0
tensorflow-hub==0.14.0
timm==1.0.14
torch==2.2.2
torchvision==0.17.2
onnxruntime==1.22.1
tqdm==4.67.1
sentencepiece==0.2.0
accelerate==1.8.1
image2base64==0.0.5
pytorch_lightning==2.5.0.post0
braceexpand==0.1.7
easydict==1.13
webdataset==0.2.111
albumentations==0.5.2
gradio_client==1.10.3
diffusers==0.30.2
hf-xet==1.1.5
huggingface-hub==0.33.1
transformers==4.53.0
uuid==1.30
safetensors==0.4.3
